feat: new theming api
