+++ src/app.html
@@ -1,15 +1,15 @@
 <!DOCTYPE html>
-<html lang="en" theme="g100">
+<html lang="en" theme="g100" data-theme="g100">
 	<head>
 		<meta charset="utf-8" />
+		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
 		<link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
 		<link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
 		<link rel="manifest" href="/site.webmanifest" />
 		<link rel="mask-icon" href="/safari-pinned-tab.svg" color="#ff3e00" />
-		<meta name="description" content="Documentation for humans.">
+		<meta name="description" content="Documentation for humans." />
 		<meta name="theme-color" content="#ff3e00" />
-		<meta name="viewport" content="width=device-width, initial-scale=1" />
 		<meta name="apple-mobile-web-app-status-bar-style" content="black-translucent" />
 		<meta name="apple-mobile-web-app-capable" content="yes" />
 		<meta name="og:title" content="SveltePress" />
