+++ src/lib/SveltePress/components/sidebar.svelte
@@ -1,7 +1,7 @@
 <script>
 	import { page, session } from '$app/stores';
 
-	import Sidebar from '$lib/SveltePress/theme/sidebar.svelte';
+	import Sidebar from '$lib/SveltePress/theme/components/sidebar.svelte';
 
 	$: parent = $page.params.group?.split('/');
 	$: guide = parent ? $session.get('sidebar').get(parent[0]) : null;
