+++ src/lib/SveltePress/theme/components/search.svelte
