+++ src/lib/SveltePress/theme/meta/package.json
@@ -1,9 +1,13 @@
 {
+	"name": "sveltepress-carbon",
+	"version": "1.0.0",
+	"type": "module",
 	"devDependencies": {
 		"@carbon/themes": "10.40.0",
 		"carbon-components": "10.41.0",
 		"carbon-components-svelte": "0.42.2",
 		"carbon-icons-svelte": "10.36.0",
-		"carbon-preprocess-svelte": "0.6.0"
+		"carbon-preprocess-svelte": "0.6.0",
+		"svelte-typewriter": "2.4.5"
 	}
 }
