--- src/lib/SveltePress/theme/styles/patches.css
@@ -1,77 +1,0 @@
-.sp--sidebar-grandparent {
-    pointer-events: none;
-}
-.bx--btn {
-    overflow-wrap: anywhere;
-    /* Safari & Opera */
-    word-break: break-word;
-}
-.bx--switcher__item-link:focus {
-    outline-color: var(--cds-button-primary);
-}
-#main-content ul {
-    list-style-type: square;
-}
-li {
-    line-height: var(--cds-body-long-02-line-height, 1.5);
-}
-a {
-    color: var(--cds-button-primary);
-    line-height: var(--cds-body-long-02-line-height, 1.5);
-}
-blockquote {
-    background-color: var(--cds-ui-background);
-    padding: 0.2rem 1rem;
-    margin: 1rem 0;
-}
-p > code {
-    background-color: var(--cds-ui-background);
-}
-p > img {
-    max-width: 100%;
-}
-pre {
-    background-color: var(--cds-ui-background);
-    margin: 1.2rem 0;
-    padding: 1rem;
-}
-:root[theme='g10'] pre,
-:root[theme='white'] pre {
-    background-color: var(--cds-ui-03);
-}
-pre,
-code {
-    white-space: pre-wrap;
-}
-.bx--btn--primary:hover {
-    background-color: var(--cds-button-primary-hover);
-}
-.bx--side-nav ~ .bx--content {
-    margin-left: 0;
-}
-:root,
-:root[theme='g10'],
-:root[theme='g100'],
-:root[theme='g90'],
-:root[theme='white'] {
-    --cds-body-long-02-line-height: 1.8;
-    --cds-interactive-01: #ff3e00;
-    --cds-button-primary: #ff3e00;
-    --cds-button-primary-hover: #e23900;
-    --cds-button-primary-active: #c23000;
-    --cds-focus: var(--cds-button-primary);
-}
-a.bx--side-nav__link[aria-current='page']::before,
-a.bx--side-nav__link--current::before {
-    background-color: var(--cds-button-primary);
-}
-a.bx--header__name:focus,
-.bx--header__action:focus,
-a.bx--header__menu-item:focus,
-.bx--skip-to-content:focus {
-    border-color: var(--cds-button-primary);
-}
-
-[role='search'].active {
-    outline-color: var(--cds-button-primary) !important;
-}
