+++ src/routes/__error.svelte
@@ -23,6 +23,6 @@
 	.sp--error-status-code {
 		font-size: 5rem;
 		font-weight: bold;
-		color: var(--cds-button-primary);
+		color: var(--cds-button-primary, #ff3e00);
 	}
 </style>
